requirements.txt
asgiref==3.8.1
cffi==1.17.0
cryptocode==0.1
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodomex==3.20.0
sqlparse==0.5.1
